# notes-app-backend/app/NotesApp/Domain/Models/Note.php
<?php

namespace App\NotesApp\Domain\Models;

use Illuminate\Contracts\Support\Jsonable;

class Note implements Jsonable
{

    public function toJson($options = 0)
    {
        // TODO: Implement toJson() method.
    }
}